msm-tsens: Reschedule work instead of causing uninterruptible sleep
The tsens_poll work causes uninterruptible sleep on the UI
thread and thus blocks other tasks from running due to the
msleep call.

To avoid this condition, we introduce a variable to hold
the staging state of the critical polling and reschedule the
work itself to be executed again after the same
duration as the msleep.

Additionally we switch the currently unused wq to
a singlethreaded wq and queue the tsens work onto
that wq.

Change-Id: I0bf48837310decd035d35ac78aa7d51614032b36
Signed-off-by: Alex Naidis <alex.naidis@paranoidandroid.co>
Signed-off-by: ben863 <r.budhiono@gmail.com>